Fix for report from Bruce: on newly installed phones, the recording tab would crash.

+++ src/fm/audioboo/app/RecordActivity.java
@@ -427,7 +427,9 @@ public class RecordActivity extends Activity
     }
 
     // Remove current boo.
-    mBoo.delete();
+    if (null != mBoo) {
+      mBoo.delete();
+    }
 
     // Create new empty Boo.
     mBoo = Globals.get().getBooManager().createBoo();
